+++ apps/mobile/src/app/settings/bilibili-account/_components/InputPhoneStep.tsx
@@ -1,4 +1,3 @@
-import { Icon } from '@expo/ui'
 import {
 	Host,
 	OutlinedTextField,
@@ -6,7 +5,7 @@ import {
 } from '@expo/ui/jetpack-compose'
 import { fillMaxWidth } from '@expo/ui/jetpack-compose/modifiers'
 import { StyleSheet, View } from 'react-native'
-import { HelperText, Text, useTheme } from 'react-native-paper'
+import { HelperText, Text, useTheme, Icon } from 'react-native-paper'
 
 import Button from '@/components/common/Button'
 import useTextFieldState from '@/hooks/useTextFieldState'
@@ -36,7 +35,7 @@ export default function InputPhoneStep({
 		<View style={styles.container}>
 			<View style={styles.header}>
 				<Icon
-					name='television-play'
+					source='television-play'
 					size={56}
 					color='#FB7299'
 				/>